# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Couture POS API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Variant"
            }
          }
        },
        "required": [
          "id",
          "name",
          "categoryId",
          "category",
          "createdAt",
          "variants"
        ]
      },
      "Variant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "barcode": {
            "type": [
              "string",
              "null"
            ]
          },
          "unitOfMeasure": {
            "$ref": "#/components/schemas/UnitOfMeasure"
          },
          "size": {
            "type": [
              "string",
              "null"
            ]
          },
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "material": {
            "type": [
              "string",
              "null"
            ]
          },
          "price": {
            "type": "string"
          },
          "reorderThreshold": {
            "type": "number"
          },
          "identityLocked": {
            "type": "boolean"
          },
          "currentStock": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "productId",
          "sku",
          "barcode",
          "unitOfMeasure",
          "size",
          "color",
          "material",
          "price",
          "reorderThreshold",
          "identityLocked",
          "currentStock",
          "createdAt"
        ]
      },
      "UnitOfMeasure": {
        "type": "string",
        "enum": [
          "piece",
          "kg",
          "gram",
          "litre",
          "ml",
          "metre",
          "box",
          "pack",
          "set",
          "pair"
        ]
      },
      "CreateProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "categoryName": {
            "type": "string",
            "maxLength": 80
          },
          "variants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "sku": {
                  "type": "string",
                  "maxLength": 40,
                  "pattern": "^[A-Za-z0-9-]+$"
                },
                "barcode": {
                  "type": "string",
                  "pattern": "^\\d{8,14}$"
                },
                "unitOfMeasure": {
                  "$ref": "#/components/schemas/UnitOfMeasure"
                },
                "size": {
                  "type": "string",
                  "maxLength": 50
                },
                "color": {
                  "type": "string",
                  "maxLength": 50
                },
                "material": {
                  "type": "string",
                  "maxLength": 50
                },
                "price": {
                  "type": "number",
                  "minimum": 0
                },
                "reorderThreshold": {
                  "type": "number",
                  "minimum": 0
                }
              },
              "required": [
                "price"
              ]
            },
            "minItems": 1
          }
        },
        "required": [
          "name",
          "variants"
        ]
      },
      "UpdateVariantRequest": {
        "type": "object",
        "properties": {
          "size": {
            "type": "string",
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "maxLength": 50
          },
          "material": {
            "type": "string",
            "maxLength": 50
          },
          "barcode": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^\\d{8,14}$"
          },
          "unitOfMeasure": {
            "$ref": "#/components/schemas/UnitOfMeasure"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "reorderThreshold": {
            "type": "number",
            "minimum": 0
          }
        }
      },
...
      "SupplierProductWithVariantList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/SupplierProductWithVariant"
        }
      },
      "SupplierProductWithVariant": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SupplierProduct"
          },
          {
            "type": "object",
            "properties": {
              "sku": {
                "type": "string"
              },
              "productName": {
                "type": "string"
              },
              "size": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "color": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "material": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "sku",
              "productName",
              "size",
              "color",
              "material"
            ]
          }
        ]
      },
      "SupplierProduct": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid"
          },
          "supplierName": {
            "type": "string"
          },
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "leadTimeDays": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "unitCost": {
            "type": [
              "string",
              "null"
            ]
          },
          "supplierSku": {
            "type": [
              "string",
              "null"
            ]
          },
          "minOrderQty": {
            "type": [
              "integer",
              "null"
            ],
            "exclusiveMinimum": 0
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "supplierId",
          "supplierName",
          "variantId",
          "isPrimary",
          "leadTimeDays",
          "unitCost",
          "supplierSku",
          "minOrderQty",
          "createdAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/products": {
      "get": {
        "description": "List the caller's tenant's products with variants and current stock. Optional `search` filters to an exact-sku or partial-name match (CHECK-01/CHECK-02).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a product with one or more variants (D-01/D-02). SKU auto-generated per variant unless supplied.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "SKU collision"
          }
        }
      }
    },
    "/products/{productId}": {
      "get": {
        "description": "Get a single product with its variants and current stock.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "productId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/products/{productId}/variants/{variantId}": {
      "patch": {
        "description": "Edit a variant. Price/reorderThreshold always editable; size/color/material blocked once stock has moved (D-04).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "productId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "variantId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVariantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Variant updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Variant"
                }
              }
            }
          },
          "404": {
            "description": "Variant not found"
          },
          "409": {
            "description": "Variant identity is locked"
          }
        }
      }
    },
...
    "/suppliers/{supplierId}/products": {
      "get": {
        "description": "Every product this supplier is linked to, for the supplier detail page's Products supplied tab.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "supplierId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Products supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierProductWithVariantList"
                }
              }
            }
          },
          "404": {
            "description": "Supplier not found"
          }
        }
      }
    },
...
  },
...
}